// src/redux/types.ts
export enum ActionTypes {
    UP_FLASK = 'UP_FLASK',
    DOWN_FLASK = 'DOWN_FLASK',
    INCREMENT_NUMBER_OF_FINISHED_FLASKS = 'INCREMENT_NUMBER_OF_FINISHED_FLASKS',
    DECREMENT_NUMBER_OF_FINISHED_FLASKS = 'DECREMENT_NUMBER_OF_FINISHED_FLASKS',
    NUMBER_OF_FILLED_FLASKS = 'NUMBER_OF_FILLED_FLASKS',
    SET_COLORS_FOR_FLASKS = 'SET_COLORS_FOR_FLASKS',
    END_LEVEL = 'END_LEVEL',
    START_LEVEL = 'START_LEVEL',
    SET_TRANSFERABLE_COLORS_STYLE = 'SET_TRANSFERABLE_COLORS_STYLE',
    SPLICE_COLORS = 'SPLICE_COLORS',
    RELOAD_LEVEL = 'RELOAD_LEVEL'
}